+++ src/msunpv/const.py
@@ -1,3 +1,3 @@
 """Constants for MSunPV WebConnect library for Python."""
 
-DEFAULT_TIMEOUT = 15
+DEFAULT_TIMEOUT = 15 #: Default time out 15s